+++ spec/support/shared_context_for_module_methods_via_class_scope.rb
@@ -1,10 +1,8 @@
 # frozen_string_literal: true
 
 RSpec.shared_context "with context for module methods via scope 'class << self'" do
-  # NOTE: This use of `before(:all)` is a performance optimization that shaves
-  # minutes off of our test suite, especially in older versions of Ruby.
-  before(:all) do
-    @_module_with_memo = Module.new do
+  let(:module_with_memo) do
+    Module.new do
       class << self
         prepend MemoWise
 
@@ -15,6 +13,4 @@ RSpec.shared_context "with context for module methods via scope 'class << self'"
       end
     end
   end
-
-  let(:module_with_memo) { @_module_with_memo } # rubocop:disable RSpec/InstanceVariable
 end